requirements.txt
discord.py==1.7.3
fastapi==0.109.1
gradio==4.19.2
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.26.0
aiofiles==23.2.1
loguru==0.7.2
cryptg==0.4.0
python-socks[asyncio]==2.4.4
Flask==2.2.2